Fixed trace args functions so they work with 32-bit architectures.

+++ builtin.h
@@ -67,13 +67,13 @@ void asm_builtin_aastore (java_objectarray *a, s4 index, java_objectheader *o);
 
 #ifdef TRACE_ARGS_NUM
 #if TRACE_ARGS_NUM == 6
-void builtin_trace_args(long a0, long a1, long a2, long a3, long a4, long a5, methodinfo *method);
+void builtin_trace_args(s8 a0, s8 a1, s8 a2, s8 a3, s8 a4, s8 a5, methodinfo *method);
 #else
-void builtin_trace_args(long a0, long a1, long a2, long a3, long a4, long a5, long a6, long a7, methodinfo *method);
+void builtin_trace_args(s8 a0, s8 a1, s8 a2, s8 a3, s8 a4, s8 a5, s8 a6, s8 a7, methodinfo *method);
 #endif
 #endif
 void builtin_displaymethodstart(methodinfo *method);
-void builtin_displaymethodstop(methodinfo *method, long l, double d, float f);
+void builtin_displaymethodstop(methodinfo *method, s8 l, double d, float f);
 /* void builtin_displaymethodstop(methodinfo *method); */
 void builtin_displaymethodexception(methodinfo *method);
